+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/asset_mapper_metadata_dir.php
@@ -1,0 +1,18 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+$container->loadFromExtension('framework', [
+    'asset_mapper' => [
+        'paths' => ['assets/'],
+        'metadata_dir' => '%kernel.build_dir%/asset_mapper',
+    ],
+    'assets' => false,
+]);